update: separate pkg/logger to pkg/sentry, pkg/stdlog
