--- pkg/logger/user.go
@@ -1,6 +1,0 @@
-package logger
-
-type User struct {
-	ID   string
-	Name string
-}
